Fix leNet5
Several issues with leNet5 model fixed. It runs correctly now.
